# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ЭкоПульс API",
    "description": "API административной панели эколога/инспектора «ЭкоПульс» — триаж обращений о состоянии площадок ТКО.\n\n**Base URL:** `/api/v1`\n\n**Авторизация:** для защищённых эндпоинтов — Bearer JWT (`Authorization: Bearer <access_token>` из `POST /auth/login`). Приём фотоотчёта — заголовок `X-Intake-Token`. Отдача фото — публично.\n\n**Конверт ошибок:** `{ \"error\": { \"code\", \"message\", \"details\" } }`.\n\nРазделы, помеченные «вне мобильного API», в клиентскую документацию не входят.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Login",
        "description": "Вход по email+паролю. Access — в теле, refresh — в HttpOnly-cookie.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/regions": {
      "get": {
        "tags": [
          "Админский реестр"
        ],
        "summary": "List Regions",
        "description": "DISTINCT непустые регионы (А→Я) — наполняет дропдаун фильтра региона.",
        "operationId": "list_regions_api_v1_incidents_regions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response List Regions Api V1 Incidents Regions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/regions": {
      "get": {
        "tags": [
          "Регионы"
        ],
        "summary": "List Regions",
        "description": "Справочник регионов: фильтры search/fed + сортировка, с mno_count/incidents_count.",
        "operationId": "list_regions_api_v1_regions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "fed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fed"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "code",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegionListItem"
                  },
                  "title": "Response List Regions Api V1 Regions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Регионы"
        ],
        "summary": "Create Region",
        "description": "Добавление региона (active=true). Дубликат кода → 409.",
        "operationId": "create_region_api_v1_regions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{code}": {
      "get": {
        "tags": [
          "Регионы"
        ],
        "summary": "Get Region",
        "description": "Карточка региона по коду субъекта.",
        "operationId": "get_region_api_v1_regions__code__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Управление пользователями (вне мобильного API)"
        ],
        "summary": "Get Users",
        "description": "Список всех пользователей для экрана Настроек.",
        "operationId": "get_users_api_v1_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserListItem"
                  },
                  "type": "array",
                  "title": "Response Get Users Api V1 Users Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Управление пользователями (вне мобильного API)"
        ],
        "summary": "Create User Endpoint",
        "description": "Создаёт активного пользователя с заданным админом паролем (инвайт-флоу убран).",
        "operationId": "create_user_endpoint_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/{user_id}/password": {
      "post": {
        "tags": [
          "Управление пользователями (вне мобильного API)"
        ],
        "summary": "Set Password",
        "description": "Задаёт/сбрасывает пароль пользователю. Супер-админу — запрещено (403).",
        "operationId": "set_password_api_v1_users__user_id__password_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "delete": {
        "tags": [
          "Управление пользователями (вне мобильного API)"
        ],
        "summary": "Remove User",
        "description": "Удаляет пользователя. Нельзя удалить супер-админа, admin-роль и самого себя.",
        "operationId": "remove_user_api_v1_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "PasswordReset": {
        "properties": {
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "new_password"
        ],
        "title": "PasswordReset"
      },
...
      "RegionCreate": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "fed": {
            "type": "integer",
            "title": "Fed"
          },
          "operators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Operators"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "fed"
        ],
        "title": "RegionCreate",
        "description": "Создание региона. active=true проставляется сервисом."
      },
      "RegionDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fed": {
            "type": "integer",
            "title": "Fed"
          },
          "fed_code": {
            "type": "string",
            "title": "Fed Code"
          },
          "fed_name": {
            "type": "string",
            "title": "Fed Name"
          },
          "operators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Operators"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "last_sync": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          },
          "mno_count": {
            "type": "integer",
            "title": "Mno Count"
          },
          "incidents_count": {
            "type": "integer",
            "title": "Incidents Count"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "fed",
          "fed_code",
          "fed_name",
          "operators",
          "active",
          "mno_count",
          "incidents_count"
        ],
        "title": "RegionDetail",
        "description": "Карточка региона — те же поля, что и в списке (контракт)."
      },
      "RegionListItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fed": {
            "type": "integer",
            "title": "Fed"
          },
          "fed_code": {
            "type": "string",
            "title": "Fed Code"
          },
          "fed_name": {
            "type": "string",
            "title": "Fed Name"
          },
          "operators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Operators"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "last_sync": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          },
          "mno_count": {
            "type": "integer",
            "title": "Mno Count"
          },
          "incidents_count": {
            "type": "integer",
            "title": "Incidents Count"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "fed",
          "fed_code",
          "fed_name",
          "operators",
          "active",
          "mno_count",
          "incidents_count"
        ],
        "title": "RegionListItem",
        "description": "Строка справочника регионов. fed_code/fed_name резолвятся из FEDERAL_DISTRICTS,\nmno_count/incidents_count — подсчитываются сервисом."
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
      "UserCreate": {
        "properties": {
          "fio": {
            "type": "string",
            "title": "Fio"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ],
            "title": "Role",
            "default": "user"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "fio",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "description": "Создание пользователя админом: пароль задаётся вручную (инвайт-флоу убран)."
      },
      "UserListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "fio": {
            "type": "string",
            "title": "Fio"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ],
            "title": "Role"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "invited"
            ],
            "title": "Status"
          },
          "is_superadmin": {
            "type": "boolean",
            "title": "Is Superadmin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fio",
          "email",
          "role",
          "status",
          "is_superadmin"
        ],
        "title": "UserListItem",
        "description": "Строка списка пользователей в Настройках."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}